Tests the specification of a linear model using wild-bootstrap.

Tests the specification of a linear model using wild-bootstrap.

dominguez_lobato_test(
  model,
  distribution = "rnorm",
  statistic = "cvm_value",
  times = 300,
  quantiles = c(0.9, 0.95, 0.99),
  verbose = FALSE,
  n_cores = 1
)

Arguments

model
An lm model.

distribution
Type of noise added to residuals, ej `rnorm` or `rrademacher`.

statistic
Type of statistic to be used, can be one of `cvm_value` or `kmv_value`.

times
Number of bootstrap samples.

quantiles
Vector of quantiles to calculate pvalues.

verbose
TRUE to print each bootstrap iteration.

n_cores
Number of cores to be used.

Value

A list with dataframe results and the ordered values of each bootstrap iteration.

References

Manuel A. Dominguez and Ignacio N. Lobato (2019). Specification Testing with Estimated Variables. Econometric Reviews.

Examples

x <- 1:10
y <- 1:10
model <- lm(y~x-1)
dominguez_lobato_test(model)

#> $test
#> # A tibble: 1 x 7
#>   name_distributi… name_statistic statistic p_value quantile_90 quantile_95
#>   <chr>            <chr>              <dbl>   <dbl>       <dbl>       <dbl>
#> 1 rnorm            cvm_value       1.04e-30   0.573    2.56e-30    4.88e-30
#> # … with 1 more variable: quantile_99 <dbl>
#> 
#> $bootstrap
#>   [1] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [6] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [11] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [16] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [21] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [26] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [31] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [36] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [41] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [46] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [51] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [56] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [61] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [66] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [71] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [76] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [81] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [86] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [91] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [96] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [101] 0.000000e+00 0.000000e+00 0.000000e+00 2.755241e-33 2.755241e-33
#> [106] 2.755241e-33 2.755241e-33 2.755241e-33 2.755241e-33 4.197722e-33
#> [111] 4.197722e-33 4.197722e-33 4.197722e-33 4.197722e-33 2.586721e-32
#> [116] 2.586721e-32 2.586721e-32 2.586721e-32 2.586721e-32 2.586721e-32
#> [121] 1.495218e-31 1.865231e-31 1.865231e-31 3.829180e-31 7.659208e-31
#> [126] 7.659208e-31 7.659208e-31 7.659208e-31 1.036836e-30 1.036836e-30
#> [131] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#> [136] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#> [141] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#> [146] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#> [151] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#> [156] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#> [161] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#> [166] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#> [171] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#> [176] 1.036836e-30 1.036836e-30 1.036836e-30 1.125603e-30 1.125603e-30
#> [181] 1.125603e-30 1.125603e-30 1.139305e-30 1.139305e-30 1.139305e-30
#> [186] 1.251750e-30 1.251750e-30 1.251750e-30 1.251750e-30 1.251750e-30
#> [191] 1.251750e-30 1.273804e-30 1.273804e-30 1.273804e-30 1.273804e-30
#> [196] 1.273804e-30 1.273804e-30 1.273804e-30 1.273804e-30 1.362521e-30
#> [201] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [206] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [211] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [216] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [221] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [226] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [231] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [236] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [241] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [246] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [251] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [256] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#> [261] 1.362521e-30 1.362521e-30 1.362521e-30 1.633549e-30 1.633549e-30
#> [266] 1.736479e-30 1.736479e-30 2.063801e-30 2.063801e-30 2.494766e-30
#> [271] 3.164669e-30 3.291328e-30 4.346354e-30 4.670251e-30 4.764309e-30
#> [276] 4.764309e-30 4.764309e-30 4.764309e-30 4.764309e-30 4.764309e-30
#> [281] 4.764309e-30 4.764309e-30 4.764309e-30 4.764309e-30 4.880691e-30
#> [286] 4.880691e-30 4.880691e-30 4.880691e-30 4.880691e-30 4.880691e-30
#> [291] 4.880691e-30 4.880691e-30 4.880691e-30 4.880691e-30 4.880691e-30
#> [296] 4.880691e-30 4.880691e-30 4.980304e-30 5.166660e-30 8.134316e-30
#> 
#> attr(,"class")
#> [1] "dl_test"

dominguez_lobato_test(model, distribution = "rmammen_point", statistic = "kmv_value")

#> $test
#> # A tibble: 1 x 7
#>   name_distributi… name_statistic statistic p_value quantile_90 quantile_95
#>   <chr>            <chr>              <dbl>   <dbl>       <dbl>       <dbl>
#> 1 rmammen_point    kmv_value       3.64e-15    0.98    4.78e-15    4.78e-15
#> # … with 1 more variable: quantile_99 <dbl>
#> 
#> $bootstrap
#>   [1] 3.550866e-15 3.550866e-15 3.550866e-15 3.550866e-15 3.550866e-15
#>   [6] 3.550866e-15 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15
#>  [11] 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15
#>  [16] 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15
#>  [21] 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15
#>  [26] 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15
#>  [31] 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15 3.639438e-15
#>  [36] 3.639438e-15 3.711864e-15 3.711864e-15 3.711864e-15 3.711864e-15
#>  [41] 3.711864e-15 3.750768e-15 3.750768e-15 3.750768e-15 3.750768e-15
#>  [46] 3.800435e-15 3.800435e-15 3.800435e-15 3.800435e-15 3.800435e-15
#>  [51] 3.800435e-15 3.800435e-15 3.800435e-15 3.800435e-15 3.800435e-15
#>  [56] 3.800435e-15 3.800435e-15 3.839339e-15 3.839339e-15 3.839339e-15
#>  [61] 3.839339e-15 3.839339e-15 3.839339e-15 3.839339e-15 3.839339e-15
#>  [66] 3.839339e-15 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15
#>  [71] 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15
#>  [76] 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15
#>  [81] 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15
#>  [86] 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15
#>  [91] 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15 3.893644e-15
#>  [96] 3.893644e-15 3.893644e-15 3.911766e-15 3.911766e-15 3.983387e-15
#> [101] 3.983387e-15 3.983387e-15 3.983387e-15 3.983387e-15 3.983387e-15
#> [106] 3.983387e-15 3.983387e-15 3.983387e-15 3.983387e-15 3.983387e-15
#> [111] 3.983387e-15 3.983387e-15 3.983387e-15 4.000337e-15 4.000337e-15
#> [116] 4.000337e-15 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15
#> [121] 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15
#> [126] 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15
#> [131] 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15
#> [136] 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15
#> [141] 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15
#> [146] 4.032926e-15 4.032926e-15 4.032926e-15 4.032926e-15 4.071105e-15
#> [151] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [156] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [161] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [166] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [171] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [176] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [181] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [186] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [191] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [196] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [201] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [206] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [211] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [216] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [221] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [226] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [231] 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15 4.071105e-15
#> [236] 4.071105e-15 4.071105e-15 4.345513e-15 4.345513e-15 4.345513e-15
#> [241] 4.345513e-15 4.518690e-15 4.518690e-15 4.518690e-15 4.518690e-15
#> [246] 4.518690e-15 4.518690e-15 4.518690e-15 4.518690e-15 4.606983e-15
#> [251] 4.606983e-15 4.606983e-15 4.606983e-15 4.606983e-15 4.606983e-15
#> [256] 4.606983e-15 4.606983e-15 4.784444e-15 4.784444e-15 4.784444e-15
#> [261] 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15
#> [266] 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15
#> [271] 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15
#> [276] 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15
#> [281] 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15
#> [286] 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15
#> [291] 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15
#> [296] 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15 4.784444e-15
#> 
#> attr(,"class")
#> [1] "dl_test"

dominguez_lobato_test(model, times = 100)

#> $test
#> # A tibble: 1 x 7
#>   name_distributi… name_statistic statistic p_value quantile_90 quantile_95
#>   <chr>            <chr>              <dbl>   <dbl>       <dbl>       <dbl>
#> 1 rnorm            cvm_value       1.04e-30    0.61    1.36e-30    4.77e-30
#> # … with 1 more variable: quantile_99 <dbl>
#> 
#> $bootstrap
#>   [1] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [6] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [11] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [16] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [21] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [26] 0.000000e+00 0.000000e+00 0.000000e+00 2.755241e-33 4.197722e-33
#>  [31] 4.197722e-33 4.197722e-33 2.586721e-32 2.586721e-32 1.865231e-31
#>  [36] 1.865231e-31 2.319198e-31 3.237446e-31 7.659208e-31 1.036836e-30
#>  [41] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#>  [46] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#>  [51] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#>  [56] 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30 1.036836e-30
#>  [61] 1.036836e-30 1.125603e-30 1.125603e-30 1.229209e-30 1.251750e-30
#>  [66] 1.251750e-30 1.251750e-30 1.362521e-30 1.362521e-30 1.362521e-30
#>  [71] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#>  [76] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#>  [81] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#>  [86] 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30 1.362521e-30
#>  [91] 1.362521e-30 1.736479e-30 4.145302e-30 4.764309e-30 4.764309e-30
#>  [96] 4.880691e-30 4.880691e-30 4.880691e-30 4.880691e-30 1.078146e-29
#> 
#> attr(,"class")
#> [1] "dl_test"